# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Inventory Management API
  version: 1.0.0
  description: A REST API for managing personal inventory with user authentication, items, categories, and locations.
  contact:
    name: API Support
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://inventory-api.onrender.com
    description: Production server on Render
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Item:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        quantity:
          type: integer
          minimum: 1
        createdAt:
          type: string
          format: date-time
        userId:
          type: integer
        categoryId:
          type: integer
        locationId:
          type: integer
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        userId:
          type: integer
    Location:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        userId:
          type: integer
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
paths:
...
  /api/items:
    get:
      summary: Get all items
      description: Retrieve all inventory items (public endpoint)
      tags:
        - Items
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by item name or description
        - name: sort
          in: query
          schema:
            type: string
            enum: [name:asc, name:desc, quantity:asc, quantity:desc, createdAt:asc, createdAt:desc]
          description: Sort items
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 100
          description: Items per page
      responses:
        '200':
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'

    post:
      summary: Create a new item
      description: Add a new item to inventory (requires authentication)
      tags:
        - Items
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - categoryId
                - locationId
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                  example: USB-C Hub
                description:
                  type: string
                  maxLength: 1000
                  example: 7-port hub with HDMI out
                quantity:
                  type: integer
                  minimum: 1
                  default: 1
                  example: 1
                categoryId:
                  type: integer
                  example: 1
                locationId:
                  type: integer
                  example: 1
      responses:
        '201':
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: Missing or invalid fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No valid JWT provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/categories:
    get:
      summary: Get all categories
      description: Retrieve all categories (public endpoint)
      tags:
        - Categories
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'

    post:
      summary: Create a new category
      description: Add a new category (requires authentication)
      tags:
        - Categories
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                  example: Electronics
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Missing or invalid fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No valid JWT provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/locations:
    get:
      summary: Get all locations
      description: Retrieve all locations (public endpoint)
      tags:
        - Locations
      responses:
        '200':
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Location'

    post:
      summary: Create a new location
      description: Add a new location (requires authentication)
      tags:
        - Locations
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                  example: Garage Shelf
      responses:
        '201':
          description: Location created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          description: Missing or invalid fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No valid JWT provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...